+++ kmtool/db/postgresql/schema.sql
@@ -19,3 +19,17 @@ CREATE TABLE km_asserted_attribute (
 );
 
 CREATE INDEX km_asserted_attribute_eppn ON km_asserted_attribute (eppn);
+
+-- Table for people whose IdPs don't send email to self assert their address
+-- See kmnoemail.php
+
+DROP TABLE IF EXISTS km_email_confirm;
+
+CREATE TABLE km_email_confirm (
+    id SERIAL PRIMARY KEY,
+    eppn VARCHAR NOT NULL,
+    email VARCHAR NOT NULL,
+    nonce VARCHAR NOT NULL,
+    created timestamp DEFAULT (NOW() at time zone 'utc') NOT NULL
+);
+